Move username and password to their own folder
