+++ waspc/data/Generator/templates/sdk/wasp/auth/username/actions/login.ts
@@ -1,8 +1,8 @@
 {{={= =}=}}
 import { api, handleApiError } from 'wasp/client/api'
-import { initSession } from './helpers/user'
+import { initSession } from '../../helpers/user'
 
-export default async function login(data: { username: string, password: string }): Promise<void> {
+export async function login(data: { username: string, password: string }): Promise<void> {
   try {
     const response = await api.post('{= loginPath =}', data)
