+++ waspc/data/Generator/templates/sdk/wasp/auth/username/actions/signup.ts
@@ -1,7 +1,7 @@
 {{={= =}=}}
 import { api, handleApiError } from 'wasp/client/api'
 
-export default async function signup(data: { username: string; password: string }): Promise<void> {
+export async function signup(data: { username: string; password: string }): Promise<void> {
   try {
     await api.post('{= signupPath =}', data)
   } catch (error) {
