+++ waspc/data/Generator/templates/sdk/wasp/auth/username/index.ts
@@ -1,0 +1,3 @@
+export { login } from './actions/login'
+export { signup } from './actions/signup'
+
