+++ waspc/data/Generator/templates/sdk/wasp/client/auth/username.ts
@@ -1,2 +1,2 @@
-export { default as login } from '../../auth/login'
-export { default as signup } from '../../auth/signup'
+export { login, signup } from '../../auth/username'
+
